+++ Airship/AirshipCore/Source/AirshipVersion.swift
@@ -5,7 +5,7 @@ import Foundation;
 
 @objc(UAirshipVersion)
 public class AirshipVersion : NSObject {
-    public static let version = "16.9.4"
+    public static let version = "16.10.0-beta"
 
     @objc
     public class func get() -> String {