+++ Airship/AirshipCore/Tests/TestChannelAudienceManager.swift
@@ -5,6 +5,7 @@ import AirshipCore
 
 @objc(UATestChannelAudienceManager)
 public class TestChannelAudienceManager : NSObject, ChannelAudienceManagerProtocol {
+
     public var contactUpdates: [SubscriptionListUpdate] = []
     
     public var pendingAttributeUpdates: [AttributeUpdate] = []
@@ -46,5 +47,8 @@ public class TestChannelAudienceManager : NSObject, ChannelAudienceManagerProtoc
 
     public func processContactSubscriptionUpdates(_ updates: [SubscriptionListUpdate]) {
         self.contactUpdates.append(contentsOf: updates)
+    }
+
+    public func addLiveActivityUpdate(_ update: LiveActivityUpdate) {
     }
 }