+++ Airship/AirshipCore/Tests/TestChannelBulkUpdateAPIClient.swift
@@ -5,14 +5,18 @@ import AirshipCore
 
 class TestChannelBulkUpdateAPIClient: ChannelBulkUpdateAPIClientProtocol {
     
-    var updateCallback: ((String, [SubscriptionListUpdate]?, [TagGroupUpdate]?, [AttributeUpdate]?, ((HTTPResponse?, Error?) -> Void)) -> Void)?
+    var updateCallback: ((String, AudienceUpdate, ((HTTPResponse?, Error?) -> Void)) -> Void)?
     var defaultCallback: ((String) -> Void)?
     
     init() {}
     
-    func update(channelID: String, subscriptionListUpdates: [SubscriptionListUpdate]?, tagGroupUpdates: [TagGroupUpdate]?, attributeUpdates: [AttributeUpdate]?, completionHandler: @escaping (HTTPResponse?, Error?) -> Void) -> Disposable {
+    func update(
+        _ update: AudienceUpdate,
+        channelID: String,
+        completionHandler: @escaping (HTTPResponse?, Error?) -> Void
+    ) -> Disposable {
         if let callback = updateCallback {
-            callback(channelID, subscriptionListUpdates, tagGroupUpdates, attributeUpdates, completionHandler)
+            callback(channelID, update, completionHandler)
         } else {
             defaultCallback?("update")
         }